fix(input): resolve pointer bounds and overlay anchors in visual space
`hit_test` compensates for scroll offsets, but everything after it re-checked
"is this point inside me" against either `global_layout_rect` (content space,
off by the scroll offset) or `last_rect or global_layout_rect` (paint state,
`None` until the first paint). So a Slider inside a scrolled Scrollable was
dead, a popup()/Menu anchored in a scrolled region opened at its layout
position, and under AppHarness -- which lays out without painting -- nothing
inside a scrolled region and no clip_content Box (.clip(), .corner_radius(),
ExtendedFab) was clickable at all.

Add `WidgetKernel.global_visual_rect` (layout rect + each ancestor's
`visual_offset()`), the definition the dev bridge's
`perception.global_visual_rect` already used, and move every input-path
reader and anchor provider onto it: interaction bounds, Slider track/handle,
Scrollable viewport check, EditableText caret, popup/menu/search/menubar
anchors. `Box.hit_test`'s clip check now reads `visual_clip_rect()` (layout
state), taking `last_rect` out of the input path entirely.

Separately, `BuilderHostMixin` now descends into `_built` from
`_hit_test_children`, so `block_pointer()` / `defer_pointer()` over a widget
with a `build()` no longer hides its whole content.

Closes #660

+++ src/nuiitivet/overlay/overlay_position.py
@@ -134,7 +134,7 @@ class OverlayPosition:
 
     @staticmethod
     def anchored(
-        rect_provider: Callable[[], Optional[Tuple[int, int, int, int]]],
+        rect_provider: Callable[[], Optional[Tuple[float, float, float, float]]],
         target_anchor: AlignmentLike = "bottom-left",
         content_anchor: AlignmentLike = "top-left",
         offset: Tuple[float, float] = (0.0, 0.0),
@@ -339,7 +339,7 @@ class _AnchoredPositionedContent(Widget):
         self,
         child: Widget,
         *,
-        rect_provider: Callable[[], Optional[Tuple[int, int, int, int]]],
+        rect_provider: Callable[[], Optional[Tuple[float, float, float, float]]],
         target_anchor: Tuple[str, str],
         content_anchor: Tuple[str, str],
         offset: Tuple[float, float],
@@ -420,7 +420,7 @@ class _AnchoredPositionedContent(Widget):
 
     def _place(
         self,
-        rect: Tuple[int, int, int, int],
+        rect: Tuple[float, float, float, float],
         cw: int,
         ch: int,
         target: Tuple[str, str],
@@ -459,7 +459,7 @@ class _AnchoredOverlayPosition(OverlayPosition):
 
     def __init__(
         self,
-        rect_provider: Callable[[], Optional[Tuple[int, int, int, int]]],
+        rect_provider: Callable[[], Optional[Tuple[float, float, float, float]]],
         target_anchor: AlignmentLike,
         content_anchor: AlignmentLike,
         offset: Tuple[float, float],
